Fix the command handling of NM
