+++ extension/bridge/src/main.rs
@@ -13,6 +13,9 @@ use std::sync::Arc;
 use std::{thread, time};
 use std::env;
 
+mod commands;
+mod io;
+
 const SERVER_AND_PORT: &str = "127.0.0.1:8754";
 
 const ALLOW_LISTED_WEBEXTENSIONS: [&str;2] = [
@@ -20,116 +23,6 @@ const ALLOW_LISTED_WEBEXTENSIONS: [&str;2] = [
     "vpn@mozilla.com"
 ];
 
-#[derive(PartialEq)]
-enum ReaderState {
-    ReadingLength,
-    ReadingBuffer,
-}
-
-pub struct Reader {
-    state: ReaderState,
-    buffer: Vec<u8>,
-    length: usize,
-}
-
-impl Reader {
-    pub fn new() -> Reader {
-        Reader {
-            state: ReaderState::ReadingLength,
-            buffer: Vec::new(),
-            length: 0,
-        }
-    }
-
-    pub fn read_input<R: Read>(&mut self, mut input: R) -> Option<Value> {
-        // Until we are able to read things from the stream...
-        loop {
-            if self.state == ReaderState::ReadingLength {
-                assert!(self.buffer.len() < size_of::<u32>());
-
-                let mut buffer = vec![0; size_of::<u32>() - self.buffer.len()];
-                match input.read(&mut buffer) {
-                    Ok(size) => {
-                        // Maybe we have read just part of the buffer. Let's append
-                        // only what we have been read.
-                        buffer.truncate(size);
-                        self.buffer.append(&mut buffer);
-
-                        // Not enough data yet.
-                        if self.buffer.len() < size_of::<u32>() {
-                            continue;
-                        }
-
-                        // Let's convert our buffer into a u32.
-                        let mut rdr = Cursor::new(&self.buffer);
-                        self.length = rdr.read_u32::<NativeEndian>().unwrap() as usize;
-                        if self.length == 0 {
-                            continue;
-                        }
-
-                        self.state = ReaderState::ReadingBuffer;
-                        self.buffer = Vec::with_capacity(self.length);
-                    }
-                    _ => return None,
-                }
-            }
-
-            if self.state == ReaderState::ReadingBuffer {
-                assert!(self.length > 0);
-                assert!(self.buffer.len() < self.length);
-
-                let mut buffer = vec![0; self.length - self.buffer.len()];
-                match input.read(&mut buffer) {
-                    Ok(size) => {
-                        // Maybe we have read just part of the buffer. Let's append
-                        // only what we have been read.
-                        buffer.truncate(size);
-                        self.buffer.append(&mut buffer);
-
-                        // Not enough data yet.
-                        if self.buffer.len() < self.length {
-                            continue;
-                        }
-
-                        match serde_json::from_slice(&self.buffer) {
-                            Ok(value) => {
-                                self.buffer.clear();
-                                self.state = ReaderState::ReadingLength;
-                                return Some(value);
-                            }
-                            _ => {
-                                self.buffer.clear();
-                                self.state = ReaderState::ReadingLength;
-                                continue;
-                            }
-                        }
-                    }
-                    _ => return None,
-                }
-            }
-        }
-    }
-}
-
-fn write_output<W: Write>(mut output: W, value: &Value) -> Result<(), std::io::Error> {
-    let msg = serde_json::to_string(value)?;
-    let len = msg.len();
-    output.write_u32::<NativeEndian>(len as u32)?;
-    output.write_all(msg.as_bytes())?;
-    output.flush()?;
-    Ok(())
-}
-
-fn write_vpn_down(error: bool) {
-    let field = if error { "error" } else { "status" };
-    let value = json!({field: "vpn-client-down"});
-    write_output(std::io::stdout(), &value).expect("Unable to write to STDOUT");
-}
-
-fn write_vpn_up() {
-    let value = json!({"status": "vpn-client-up"});
-    write_output(std::io::stdout(), &value).expect("Unable to write to STDOUT");
-}
 
 fn main() {
     /*
@@ -168,24 +61,23 @@ fn main() {
     let waker = Arc::new(Waker::new(poll.registry(), STDIN_WAKER).unwrap());
     let waker_cloned = waker.clone();
     thread::spawn(move || {
-        let mut r = Reader::new();
+        let mut r = io::Reader::new();
         loop {
             match r.read_input(std::io::stdin()) {
                 Some(value) => {
-                    if value == "bridge_ping" {
-                        // A simple ping/pong message.
-                        let pong_value = json!("bridge_pong");
-                        write_output(std::io::stdout(), &pong_value)
-                            .expect("Unable to write to STDOUT");
-                    } else {
-                        // For a real message, we wake up the main thread.
-                        sender
-                            .send(value)
-                            .expect("Unable to send data to the main thread");
-                        waker_cloned.wake().expect("Unable to wake the main thread");
+                    match commands::handle(&value){
+                        Ok(true) =>{
+                            // Command was handled successfully.
+                        }
+                        _ =>{
+                            // For a real message, we wake up the main thread.
+                            sender
+                                .send(value)
+                                .expect("Unable to send data to the main thread");
+                            waker_cloned.wake().expect("Unable to wake the main thread");
+                        }
                     }
                 }
-
                 None => {
                     thread::sleep(time::Duration::from_millis(500));
                 }
@@ -204,7 +96,7 @@ fn main() {
                     .register(&mut stream, VPN, Interest::READABLE | Interest::WRITABLE)
                     .unwrap();
 
-                let mut r = Reader::new();
+                let mut r = io::Reader::new();
 
                 // This second loop processes messages coming from the tcp stream and from the
                 // STDIN thread via the waker/sender.
@@ -221,7 +113,7 @@ fn main() {
                                 {
                                     poll.registry().deregister(&mut stream).unwrap();
                                     vpn_connected = false;
-                                    write_vpn_down(false);
+                                    io::write_vpn_down(false);
                                     thread::sleep(time::Duration::from_millis(500));
                                     new_connection_needed = true;
                                 }
@@ -233,7 +125,7 @@ fn main() {
 
                                 if !vpn_connected && (event.is_readable() || event.is_writable()) {
                                     vpn_connected = true;
-                                    write_vpn_up();
+                                    io::write_vpn_up();
                                 }
 
                                 if event.is_readable() {
@@ -241,7 +133,7 @@ fn main() {
                                     loop {
                                         match r.read_input(&mut stream) {
                                             Some(value) => {
-                                                write_output(std::io::stdout(), &value)
+                                                io::write_output(std::io::stdout(), &value)
                                                     .expect("Unable to write to STDOUT");
                                             }
                                             _ => {
@@ -253,8 +145,8 @@ fn main() {
                             }
                             STDIN_WAKER => {
                                 let value = receiver.recv().unwrap();
-                                if let Err(_) = write_output(&mut stream, &value) {
-                                    write_vpn_down(true);
+                                if let Err(_) = io::write_output(&mut stream, &value) {
+                                    io::write_vpn_down(true);
                                 }
                             }
                             _ => unreachable!(),
